md/raid5: Fix possible confusion when multiple write errors occur.
commit 04da71a39dda6a94bfbaed895ddda047fde2d0ee
    md: raid5 crash during degradation

Fixed a crash in an overly simplistic way which could leave
R5_WriteError or R5_MadeGood set in the stripe cache for devices
for which it is no longer relevant.
When those devices are removed and spares added the flags are still
set and can cause incorrect behaviour.

commit 4197e92e0f43bccf660f0f42c9a49798e63334de
    md/raid5: preferentially read from replacement device if possible.

Fixed the same bug if a more effective way, so we can now revert
the original commit.

Reported-and-tested-by: Alexander Lyakas <alex.bolshoy@gmail.com>
Cc: stable@vger.kernel.org (3.2+ - 3.2 will need a different fix though)
Fixes: 04da71a39dda6a94bfbaed895ddda047fde2d0ee
Signed-off-by: NeilBrown <neilb@suse.de>